requirements.txt
polars==0.20.18
XlsxWriter==3.2.0
xlsx2csv==0.8.2
apache-airflow-providers-postgres==5.4.0
psycopg2-binary==2.9.6